requirements.txt
pandas>=2.2
streamlit>=1.34
google-generativeai>=0.4
openpyxl>=3.1   # Excel read/write backend
python-dotenv>=1.0
